Use `capybara-screenshot` again (#3782)
* Use `capybara-screenshot` again

A while ago I added a rails monkeypatch so that we can get HTML
screenshots on spec failures. The idea was that the patch would be
incorporated into Rails and we could eventually get rid of the
monkeypatch. However, it doesn't look like this will happen.

capybara-screenshot works just fine, and although it requires a small
monkeypatch too, it's much more maintainable than the previous version.
So I'd rather go back to it, since the current monkeypatch is too big
and specific and might cause trouble in the future.